requirements.txt
cloudinary==1.22.0
Django==3.0.7
django-bootstrap4==2.2.0
django-model-helpers==1.3.2
factory-boy==3.0.1
Faker==4.1.0
Pillow==7.1.2
psycopg2==2.8.5
pytest==5.4.3
pytest-django==3.9.0
python-decouple==3.3
python-magic==0.4.18
